Adding iam policy and moving everything to app.py
