+++ .chalice/config.json
@@ -3,7 +3,9 @@
   "app_name": "restaurant-journal",
   "stages": {
     "dev": {
-      "api_gateway_stage": "api"
+      "api_gateway_stage": "api",
+      "autogen_policy": false,
+      "iam_policy_file": "policy.json"
     }
   }
 }
