+++ .chalice/deployed/dev.json
@@ -1,10 +1,10 @@
 {
   "resources": [
     {
-      "name": "default-role",
+      "name": "api_handler_role",
       "resource_type": "iam_role",
-      "role_arn": "arn:aws:iam::132981671275:role/restaurant-journal-dev",
-      "role_name": "restaurant-journal-dev"
+      "role_arn": "arn:aws:iam::132981671275:role/restaurant-journal-dev-api_handler",
+      "role_name": "restaurant-journal-dev-api_handler"
     },
     {
       "name": "api_handler",
